Move tests related to the scheduler into their own module
